+++ packages/app/hivecommand-backend/src/schema/operations/index.ts
@@ -1,6 +1,7 @@
 import { nanoid } from "nanoid";
 import { MQTTClient, MQTTPublisher } from '@hive-command/amqp-client'
 import { PrismaClient } from "@hive-command/data";
+import { subject } from "@casl/ability";
 
 export default (prisma: PrismaClient, deviceChannel?: MQTTPublisher) => {
 
@@ -27,19 +28,22 @@ export default (prisma: PrismaClient, deviceChannel?: MQTTPublisher) => {
 					}
 				})
 
-				const program = await prisma.program.findFirst({
-					where: {
-						usedBy: {
-							some: {
-								id: args.deviceId
-							}
-						}
-					},
-				})
+				if(!device) return new Error("No device found")
+				if(!context?.jwt?.acl?.can('control', subject('CommandDevice', device))) throw new Error('Not allowed to control this device');
+
+
+				// const program = await prisma.program.findFirst({
+				// 	where: {
+				// 		usedBy: {
+				// 			some: {
+				// 				id: args.deviceId
+				// 			}
+				// 		}
+				// 	},
+				// })
 		
 				// const dataType = program?.devices?.find((a) => `${a.type?.tagPrefix ? a.type?.tagPrefix : ''}${a.tag}` == args.deviceName)?.type?.state?.find((a) => a.key == args.key)?.type;
 
-				if(!device) return new Error("No device found")
 
 				// let stateChange = {
 				// 	address: `opc.tcp://${device?.network_name}:8440`,
@@ -47,6 +51,16 @@ export default (prisma: PrismaClient, deviceChannel?: MQTTPublisher) => {
 				// 	dataType,
 				// 	value: args.value
 				// }
+
+				await prisma.deviceValueChange.create({
+					data: {
+						userId: context?.jwt?.id,
+						key: args.key,
+						placeholder: args.deviceName,
+						deviceId: args.deviceId,
+						value: args.value
+					}
+				})
 
 				// return await channel.sendToQueue(`COMMAND:DEVICE:VALUE`, Buffer.from(JSON.stringify(stateChange)))
 				// await deviceChannel.channel?.assertQueue(`device:${device?.network_name}`)