+++ packages/core/command-data/prisma/migrations/20230707010410_add_device_value_change/migration.sql
@@ -1,0 +1,15 @@
+-- CreateTable
+CREATE TABLE "DeviceValueChange" (
+    "id" SERIAL NOT NULL,
+    "placeholder" TEXT NOT NULL,
+    "key" TEXT,
+    "value" TEXT NOT NULL,
+    "userId" TEXT NOT NULL,
+    "lastUpdated" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "deviceId" TEXT NOT NULL,
+
+    CONSTRAINT "DeviceValueChange_pkey" PRIMARY KEY ("id","lastUpdated")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "DeviceValueChange_deviceId_placeholder_key_userId_lastUpdat_key" ON "DeviceValueChange"("deviceId", "placeholder", "key", "userId", "lastUpdated");