+++ packages/drivers/command-ethernet-ip/src/index.ts
@@ -59,7 +59,7 @@ export default class EthernetIPDriver extends BaseCommandDriver {
         // this.controller.PLC.ta
     }
 
-    async start() {
+    async start(noRetry = false) {
         try {
             await this.controller.connect()
 
@@ -72,6 +72,15 @@ export default class EthernetIPDriver extends BaseCommandDriver {
 
         } catch (e) {
             console.error("Error starting driver", e);
+
+            if(!noRetry){
+                console.log("Retrying start in 5 seconds...")
+                await new Promise(async (resolve) => {
+                    setTimeout(() => resolve(true), 5 * 1000)
+                })
+                
+                await this.start();
+            }
         }
     }
 
@@ -226,7 +235,11 @@ export default class EthernetIPDriver extends BaseCommandDriver {
             write = (value: any) => {
                 if(tag){
 
-                    try{value = parseFloat(value)}catch(e){}
+                    try{
+                        if(!Number.isNaN(parseFloat(value))){
+                            value = parseFloat(value)
+                        }
+                    }catch(e){}
 
                     tag.value = value;
                     this.writeTag(tag);